Moves SectionTitle component to shared
