+++ src/components/Landing/Landing.js
@@ -3,7 +3,7 @@ import StyledLink from '../shared/Link';
 import Block from '../shared/Block';
 import ContentSection from '../shared/ContentSection';
 import TrackCard from './trackCard';
-import SectionTitle from '../SectionTitle';
+import SectionTitle from '../shared/SectionTitle';
 import Text from '../shared/Text';
 import { getIcon } from '../../utils';
 import './Landing.scss';
