+++ src/components/Landing/trackCard.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Link } from 'gatsby';
-import SectionTitle from '../SectionTitle';
+import SectionTitle from '../shared/SectionTitle';
 import Text from '../shared/Text';
 import Icon from '../shared/Icon';
 import './trackCard.scss';
