+++ src/components/shared/SectionTitle/SectionTitle.scss
