+++ src/components/shared/SectionTitle/index.js
